+++ api/routes/users/utils.js
@@ -1,8 +1,7 @@
 //librería para encriptación
-
 const CryptoJS = require('crypto-js')
-const Institution = require('../models/Institution')
-const Profile = require('../models/Profiles')
+const Institution = require('../../models/Institution')
+const Profile = require('../../models/Profiles')
 
 //usarlo en el create del usuario, pasarle su pass de body
 //y en el log in para chequear el mismo con lo que ya estará en db del user
@@ -10,27 +9,6 @@ const encrypt = (pass) => {
   var crypted = CryptoJS.SHA3(pass, { outputLength: 224 })
   crypted = crypted.toString()
   return crypted
-}
-
-// Nueva institucion
-async function postInstitution({ name, email, password }) {
-  try {
-    const crypted = encrypt(password)
-    console.log(name)
-    // let userOfInstitution = await Profile.find({ institution: name });
-
-    let newInstitution = await new Institution({
-      name,
-      email,
-      password: crypted,
-      // users: userOfInstitution,
-    })
-    newInstitution.save()
-
-    return newInstitution
-  } catch (error) {
-    throw new Error(error)
-  }
 }
 
 //Mesas y asignacion de usuarios
@@ -42,7 +20,6 @@ function shuffle(array) {
 }
 
 // Asigna numero de mesa al usuario y lo actualiza en la db
-
 async function asignTable(users) {
   var contador = 0
   let numTable = 1
@@ -68,7 +45,6 @@ async function asignTable(users) {
 
 module.exports = {
   encrypt,
-  postInstitution,
   shuffle,
   asignTable,
 }
